+++ test_vectortiles/test_app/models.py
@@ -1,8 +1,17 @@
 from django.contrib.gis.db import models
+from django.contrib.postgres.indexes import GinIndex
+from django.core.serializers.json import DjangoJSONEncoder
 
 
 class Layer(models.Model):
-    name = models.CharField(max_length=250)
+    name = models.CharField(max_length=250, unique=True)
+    attribution = models.CharField(max_length=250, default="", blank=True)
+    description = models.TextField(blank=True)
+    min_zoom = models.PositiveSmallIntegerField(default=0)
+    max_zoom = models.PositiveSmallIntegerField(default=22)
+
+    def __str__(self):
+        return self.name
 
 
 class Feature(models.Model):
@@ -15,3 +24,38 @@ class Feature(models.Model):
 
     class Meta:
         ordering = ("id",)
+
+
+class FullDataLayer(models.Model):
+    name = models.CharField(max_length=250, unique=True)
+    attribution = models.CharField(max_length=250, default="", blank=True)
+    description = models.TextField(blank=True)
+    include_in_tilejson = models.BooleanField(default=False)
+    min_zoom = models.PositiveSmallIntegerField(default=0)
+    max_zoom = models.PositiveSmallIntegerField(default=22)
+    update_datetime = models.DateTimeField(auto_now=True, db_index=True)
+
+    def __str__(self):
+        return self.name
+
+    class Meta:
+        ordering = ("name",)
+
+
+class FullDataFeature(models.Model):
+    geom = models.GeometryField(srid=4326)
+    layer = models.ForeignKey(
+        FullDataLayer,
+        on_delete=models.CASCADE,
+        related_name="features",
+        null=True,
+        blank=True,
+    )
+    properties = models.JSONField(default=dict, encoder=DjangoJSONEncoder)
+
+    class Meta:
+        ordering = ("id",)
+        indexes = [
+            GinIndex(fields=["properties"], name="feature_properties_gin"),
+            models.Index(models.F("properties__nature"), name="properties_nature"),
+        ]