+++ vectortiles/mixins.py
@@ -1,189 +1,182 @@
-from urllib.parse import unquote
+from urllib.parse import unquote, urljoin
 
-import mercantile
-from django.http import HttpResponse
+from django.urls import path
 
 from vectortiles import settings as app_settings
 
 
-class BaseVectorTileMixin:
-    """
-    Base Mixin to handle vector tile generation
-    """
+class BaseVectorView:
+    layer_classes = None
+    layers = None
+    prefix_tiles_url = "tiles"
 
-    vector_tile_content_type = "application/x-protobuf"
-    vector_tile_queryset = None
-    vector_tile_queryset_limit = None
-    vector_tile_layer_name = None  # name for data layer in vector tile
-    vector_tile_geom_name = "geom"  # geom field to consider in qs
-    vector_tile_fields = None  # other fields to include from qs
-    vector_tile_generation = (
-        None  # use mapbox if you installed [mapbox] sub-dependencies
-    )
-    vector_tile_extent = 4096  # define tile extent
-    vector_tile_buffer = (
-        256  # define buffer around tiles (intersected polygon display without borders)
-    )
-    vector_tile_clip_geom = (
-        True  # define if feature geometries should be clipped in tile
-    )
+    def get_layer_classes(self):
+        return self.layer_classes or []
 
-    @classmethod
-    def get_bounds(cls, x, y, z):
-        """
-        Get extent from xyz tile extent to 3857
+    def get_layer_class_kwargs(self):
+        return {}
 
-        :param x: longitude coordinate tile
-        :type x: int
-        :param y: latitude coordinate tile
-        :type y: int
-        :param z: zoom level
-        :type z: int
-
-        :return: xmin, ymin, xmax, ymax in 3857 coordinate system
-        :rtype: tuple
-        """
-        return mercantile.xy_bounds(x, y, z)
-
-    def get_vector_tile_queryset(self):
-        """Get feature queryset in tile dynamically"""
+    def get_layers(self):
         return (
-            self.vector_tile_queryset
-            if self.vector_tile_queryset is not None
-            else self.get_queryset()
+            self.layers
+            if self.layers
+            else [
+                layer_class(**self.get_layer_class_kwargs())
+                for layer_class in self.get_layer_classes()
+            ]
         )
 
-    def get_vector_tile_queryset_limit(self):
-        """Get feature limit by tile dynamically"""
-        return self.vector_tile_queryset_limit
 
-    def get_vector_tile_layer_name(self):
-        """Get layer name in tile dynamically"""
-        return self.vector_tile_layer_name
+class BaseTileJSONView(BaseVectorView):
+    # https://github.com/mapbox/tilejson-spec/tree/master/3.0.0
+    name = None
+    attribution = None
+    description = None
+    min_zoom = None  # 0 - 30 range. lower than equal than max_zoom. By default, it's min_zoom of all layers.
+    max_zoom = None  # 0 - 30 range. greater than equal than min_zoom. By default, it's max_zoom of all layers.
+    fill_zoom = None
+    legend = None
+    bounds = [-180, -85.05112877980659, 180, 85.0511287798066]
+    center = None
+    scheme = "xyz"
+    version = "1.0.0"
 
-    def get_tile(self, x, y, z):
-        """
-        Generate a mapbox vector tile as bytearray
+    def __init__(self, *args, **kwargs):
+        if self.get_min_zoom() > self.get_max_zoom():
+            raise ValueError("min_zoom must be lower than equal than max_zoom")
+        if self.get_max_zoom() < self.get_min_zoom():
+            raise ValueError("max_zoom must be greater than equal than min_zoom")
+        if not (0 <= self.get_min_zoom() <= 30):
+            raise ValueError("min_zoom should be in range 0 - 30")
+        if not (0 <= self.get_max_zoom() <= 30):
+            raise ValueError("max_zoom should be in range 0 - 30")
+        super().__init__(*args, **kwargs)
 
-        :param x: longitude coordinate tile
-        :type x: int
-        :param y: latitude coordinate tile
-        :type y: int
-        :param z: zoom level
-        :type z: int
+    def get_min_zoom(self):
+        """Get tilejson minzoom from layers or self.min_zoom"""
+        try:
+            # minimum zoom level from layers
+            layers_min_zoom = min(layer.get_min_zoom() for layer in self.get_layers())
+        except ValueError:
+            # case there is no layer defined ...
+            layers_min_zoom = None
 
-        :return: Mapbox Vector Tile
-        :rtype: bytearray
-        """
-        raise NotImplementedError()
+        if self.min_zoom is not None:
+            # if defined, return max of layers_min_zoom and self.min_zoom
+            return (
+                max(layers_min_zoom, self.min_zoom)
+                if layers_min_zoom is not None
+                else self.min_zoom
+            )
+        # if self.min_zoom not defined, return layers_min_zoom or 0
+        return layers_min_zoom if layers_min_zoom is not None else 0
 
+    def get_max_zoom(self):
+        """Get tilejson maxzoom from layers or self.max_zoom"""
+        try:
+            # maximum zoom level from layers
+            layers_max_zoom = min(layer.get_max_zoom() for layer in self.get_layers())
+        except ValueError:
+            # case there is no layer defined ...
+            layers_max_zoom = None
 
-class VectorLayer:
-    vector_tile_layer_id = ""
-    vector_tile_layer_description = ""
-    vector_tile_layer_min_zoom = 0
-    vector_tile_layer_max_zoom = 22
+        if self.max_zoom is not None:
+            # if defined, return min of layers_max_zoom and self.max_zoom
+            return (
+                min(layers_max_zoom, self.max_zoom)
+                if layers_max_zoom is not None
+                else self.max_zoom
+            )
+        # if self.max_zoom not defined, return layers_max_zoom or 30
+        return layers_max_zoom if layers_max_zoom is not None else 30
 
-    def __init__(self, id_layer, description="", min_zoom=0, max_zoom=22):
-        self.vector_tile_layer_id = id_layer
-        self.vector_tile_layer_description = description
-        self.vector_tile_layer_min_zoom = min_zoom
-        self.vector_tile_layer_max_zoom = max_zoom
+    def get_fill_zoom(self):
+        return self.fill_zoom
 
-    def get_vector_tile_layer_id(self):
-        return self.vector_tile_layer_id
+    def get_attribution(self):
+        return self.attribution
 
-    def get_vector_tile_layer_description(self):
-        return self.vector_tile_layer_description
+    def get_legend(self):
+        return self.legend
 
-    def get_vector_tile_layer_min_zoom(self):
-        return self.vector_tile_layer_min_zoom
+    def get_description(self):
+        return self.description
 
-    def get_vector_tile_layer_max_zoom(self):
-        return self.vector_tile_layer_max_zoom
+    def get_name(self):
+        return self.name
 
-    def get_vector_layer(self):
-        return {
-            "id": self.get_vector_tile_layer_id(),
-            "description": self.get_vector_tile_layer_description(),
-            "fields": {},  # self.layer_fields(layer),
-            "minzoom": self.get_vector_tile_layer_min_zoom(),
-            "maxzoom": self.get_vector_tile_layer_max_zoom(),
-        }
+    def get_bounds(self):
+        return self.bounds
 
+    def get_center(self):
+        return self.center
 
-class BaseTileJSONMixin:
-    vector_tile_tilejson_name = ""
-    vector_tile_tilejson_attribution = ""
-    vector_tile_tilejson_description = ""
+    def get_scheme(self):
+        return self.scheme
 
-    def get_vector_tile_tilejson_min_zoom(self):
-        min_zoom = min(item["minzoom"] for item in self.get_vector_layers())
-        return min_zoom or 0
+    def get_version(self):
+        return self.version
 
-    def get_vector_tile_tilejson_max_zoom(self):
-        max_zoom = max(item["maxzoom"] for item in self.get_vector_layers())
-        return max_zoom or 22
-
-    def get_vector_tile_tilejson_attribution(self):
-        return self.vector_tile_tilejson_attribution
-
-    def get_vector_tile_tilejson_description(self):
-        return self.vector_tile_tilejson_description
-
-    def get_vector_tile_tilejson_name(self):
-        return self.vector_tile_tilejson_name
-
-    def get_tile_urls(self, tile_url, base_url=""):
-        # if app_settings.TERRA_TILES_HOSTNAMES:
-        #     return [
-        #         unquote(urljoin(hostname, tile_url))
-        #         for hostname in app_settings.VECTOR_TILES_HOSTNAMES
-        #     ]
-        # else:
-        return [unquote(tile_url)]
-
-    def get_vector_layers(self):
-        raise NotImplementedError(
-            """ you should implement get_vector_layers to return a VectorLayer list """
-        )
+    def get_tile_urls(self, tile_url):
+        if app_settings.VECTOR_TILES_URLS:
+            return [
+                unquote(urljoin(hostname, tile_url))
+                for hostname in app_settings.VECTOR_TILES_URLS
+            ]
+        else:
+            return [unquote(self.request.build_absolute_uri(tile_url))]
 
     def get_tilejson(self, tile_url, version="3.0.0"):
-        # https://github.com/mapbox/tilejson-spec/tree/3.0/3.0.0
+        # https://github.com/mapbox/tilejson-spec/tree/master/3.0.0
         return {
             "tilejson": version,
-            "name": self.get_vector_tile_tilejson_name(),
+            "name": self.get_name(),
+            "description": self.get_description(),
+            "legend": self.get_legend(),
+            "attribution": self.get_attribution(),
             "tiles": self.get_tile_urls(tile_url),
-            "minzoom": self.get_vector_tile_tilejson_min_zoom(),
-            "maxzoom": self.get_vector_tile_tilejson_max_zoom(),
-            # bounds
-            # center
-            "attribution": self.get_vector_tile_tilejson_attribution(),
-            "description": self.get_vector_tile_tilejson_description(),
-            "vector_layers": self.get_vector_layers(),
+            "minzoom": self.get_min_zoom(),
+            "maxzoom": self.get_max_zoom(),
+            "fillzoom": self.get_fill_zoom(),
+            "bounds": self.get_bounds(),
+            "center": self.get_center(),
+            "scheme": self.get_scheme(),
+            "version": self.get_version(),
+            "vector_layers": [
+                layer.get_tilejson_vector_layer() for layer in self.get_layers()
+            ],
         }
 
 
-class BaseVectorTileView:
-    """Base mixin to handle vector tile in a djang oView"""
+class BaseVectorTileView(BaseVectorView):
+    """Base mixin to handle vector tile in a django View"""
 
     content_type = app_settings.VECTOR_TILES_CONTENT_TYPE
 
-    def get(self, request, z, x, y):
-        """
-        Handle GET request to serve tile
+    def get_layer_tiles(self, z, x, y):
+        layers = self.get_layers()
+        if layers:
+            return b"".join(layer.get_tile(x, y, z) for layer in layers)
+        raise Exception("No layers defined")
 
-        :param request:
-        :type request: HttpRequest
-        :param x: longitude coordinate tile
-        :type x: int
-        :param y: latitude coordinate tile
-        :type y: int
-        :param z: zoom level
-        :type z: int
+    def get_content_status(self, z, x, y):
+        content = self.get_layer_tiles(z, x, y)
+        return (content, 200) if content else (content, 204)
 
-        :rtype HTTPResponse
-        """
-        content = self.get_tile(x, y, z)
-        status = 200 if content else 204
-        return HttpResponse(content, content_type=self.content_type, status=status)
+    def get_base_url(self):
+        pass
+
+    def get_default_url_pattern(self):
+        return "{z}/{x}/{y}"
+
+    def get_default_url_matrix(self):
+        pattern = self.get_default_url_pattern()
+        return f"{pattern.replace('{z}', '<int:z>').replace('{x}', '<int:x>').replace('{y}', '<int:y>')}"
+
+    def get_url(self, prefix=None, url_name=None):
+        """Generate URL to serve vector tiles with required parameters"""
+        return path(
+            f"{prefix or self.prefix_tiles_url}/{self.get_default_url_matrix()}",
+            self.as_view(),
+            name=url_name,
+        )