+++ vectortiles/postgis/views.py
@@ -1,8 +1,0 @@
-from django.views import View
-
-from vectortiles.mixins import BaseVectorTileView
-from vectortiles.postgis.mixins import PostgisBaseVectorTile
-
-
-class MVTView(BaseVectorTileView, PostgisBaseVectorTile, View):
-    pass