+++ vectortiles/tests/test_views.py
@@ -25,84 +25,11 @@ class VectorTileTestCase(VectorTileBaseTest):
     def test_num_queries_equals_one(self):
         self.maxDiff = None
         with self.assertNumQueries(1):
-            self.client.get(
-                reverse(
-                    "feature-postgis-with-manual-vector-tile-queryset", args=(0, 0, 0)
-                )
-            )
+            self.client.get(reverse("feature", args=(0, 0, 0)))
 
-    def test_mapbox_layer(self):
+    def test_features(self):
         self.maxDiff = None
-        response = self.client.get(
-            reverse("layer-mapbox", args=(self.layer.pk, 0, 0, 0))
-        )
-        self.assertEqual(response.status_code, 200)
-        content = mapbox_vector_tile.decode(response.content)
-        self.assertDictEqual(
-            content,
-            {
-                "features": {
-                    "extent": 4096,
-                    "version": 1,
-                    "features": [
-                        {
-                            "geometry": {"type": "Point", "coordinates": [2048, 2048]},
-                            "properties": {"name": "feat1"},
-                            "id": 0,
-                            "type": 1,
-                        },
-                        {
-                            "geometry": {
-                                "type": "LineString",
-                                "coordinates": [[2048, 2048], [2059, 2059]],
-                            },
-                            "properties": {"name": "feat2"},
-                            "id": 0,
-                            "type": 2,
-                        },
-                    ],
-                }
-            },
-            content,
-        )
-
-    def test_mapbox_features(self):
-        self.maxDiff = None
-        response = self.client.get(reverse("feature-mapbox", args=(0, 0, 0)))
-        self.assertEqual(response.status_code, 200)
-        content = mapbox_vector_tile.decode(response.content)
-        self.assertDictEqual(
-            content,
-            {
-                "features": {
-                    "extent": 4096,
-                    "version": 1,
-                    "features": [
-                        {
-                            "geometry": {"type": "Point", "coordinates": [2048, 2048]},
-                            "properties": {"name": "feat1"},
-                            "id": 0,
-                            "type": 1,
-                        },
-                        {
-                            "geometry": {
-                                "type": "LineString",
-                                "coordinates": [[2048, 2048], [2059, 2059]],
-                            },
-                            "properties": {"name": "feat2"},
-                            "id": 0,
-                            "type": 2,
-                        },
-                    ],
-                }
-            },
-        )
-
-    def test_postgis_layer(self):
-        self.maxDiff = None
-        response = self.client.get(
-            reverse("layer-postgis", args=(self.layer.pk, 0, 0, 0))
-        )
+        response = self.client.get(reverse("feature", args=(0, 0, 0)))
         self.assertEqual(response.status_code, 200)
         content = mapbox_vector_tile.decode(response.content)
         self.assertDictEqual(
@@ -132,105 +59,9 @@ class VectorTileTestCase(VectorTileBaseTest):
             },
         )
 
-    def test_postgis_features(self):
+    def test_features_with_filtered_date(self):
         self.maxDiff = None
-        response = self.client.get(reverse("feature-postgis", args=(0, 0, 0)))
-        self.assertEqual(response.status_code, 200)
-        content = mapbox_vector_tile.decode(response.content)
-        self.assertDictEqual(
-            content,
-            {
-                "features": {
-                    "extent": 4096,
-                    "version": 2,
-                    "features": [
-                        {
-                            "geometry": {"type": "Point", "coordinates": [2048, 2048]},
-                            "properties": {"name": "feat1"},
-                            "id": 0,
-                            "type": 1,
-                        },
-                        {
-                            "geometry": {
-                                "type": "LineString",
-                                "coordinates": [[2048, 2048], [2059, 2059]],
-                            },
-                            "properties": {"name": "feat2"},
-                            "id": 0,
-                            "type": 2,
-                        },
-                    ],
-                }
-            },
-        )
-
-    def test_postgis_drf_api_features(self):
-        self.maxDiff = None
-        response = self.client.get(reverse("feature-postgis-drf", args=(0, 0, 0)))
-        self.assertEqual(response.status_code, 200)
-        content = mapbox_vector_tile.decode(response.content)
-        self.assertDictEqual(
-            content,
-            {
-                "features": {
-                    "extent": 4096,
-                    "version": 2,
-                    "features": [
-                        {
-                            "geometry": {"type": "Point", "coordinates": [2048, 2048]},
-                            "properties": {"name": "feat1"},
-                            "id": 0,
-                            "type": 1,
-                        },
-                        {
-                            "geometry": {
-                                "type": "LineString",
-                                "coordinates": [[2048, 2048], [2059, 2059]],
-                            },
-                            "properties": {"name": "feat2"},
-                            "id": 0,
-                            "type": 2,
-                        },
-                    ],
-                }
-            },
-        )
-
-    def test_postgis_drf_viewset_features(self):
-        self.maxDiff = None
-        response = self.client.get(reverse("feature-drf-viewset-tile", args=(0, 0, 0)))
-        self.assertEqual(response.status_code, 200)
-        content = mapbox_vector_tile.decode(response.content)
-        self.assertDictEqual(
-            content,
-            {
-                "features": {
-                    "extent": 4096,
-                    "version": 2,
-                    "features": [
-                        {
-                            "geometry": {"type": "Point", "coordinates": [2048, 2048]},
-                            "properties": {"name": "feat1"},
-                            "id": 0,
-                            "type": 1,
-                        },
-                        {
-                            "geometry": {
-                                "type": "LineString",
-                                "coordinates": [[2048, 2048], [2059, 2059]],
-                            },
-                            "properties": {"name": "feat2"},
-                            "id": 0,
-                            "type": 2,
-                        },
-                    ],
-                }
-            },
-        )
-
-    def test_postgis_features_with_filtered_date(self):
-        self.maxDiff = None
-        response = self.client.get(reverse("feature-date-postgis", args=(0, 0, 0)))
+        response = self.client.get(reverse("feature-date", args=(0, 0, 0)))
         self.assertEqual(response.status_code, 200)
         content = mapbox_vector_tile.decode(response.content)
         self.assertDictEqual(
@@ -253,117 +84,29 @@ class VectorTileTestCase(VectorTileBaseTest):
 
 
 class VectorTileTileJSONTestCase(VectorTileBaseTest):
-    def test_mapbox_layer(self):
+    def test_features(self):
         self.maxDiff = None
-        response = self.client.get(
-            reverse("layer-mapbox-tilejson", args=(self.layer.pk,))
-        )
+        response = self.client.get(reverse("feature-tilejson"))
         self.assertEqual(response.status_code, 200)
         content = response.json()
         self.assertDictEqual(
             content,
             {
                 "attribution": "© JEC",
-                "description": "generated from mapbox library",
-                "maxzoom": 22,
-                "minzoom": 0,
-                "name": "Layer's features tileset",
-                "tilejson": "3.0.0",
-                "tiles": ["/layer/2/mapbox/tile/{z}/{x}/{y}"],
-                "vector_layers": [
-                    {
-                        "description": "Feature layer",
-                        "fields": {},
-                        "id": "features",
-                        "maxzoom": 22,
-                        "minzoom": 0,
-                    }
-                ],
-            },
-        )
-
-    def test_mapbox_features(self):
-        self.maxDiff = None
-        response = self.client.get(reverse("feature-mapbox-tilejson"))
-        self.assertEqual(response.status_code, 200)
-        content = response.json()
-        self.assertDictEqual(
-            content,
-            {
-                "attribution": "© JEC",
-                "description": "generated from mapbox library",
-                "maxzoom": 22,
+                "description": "feature tileset",
+                "bounds": [-180, -85.05112877980659, 180, 85.0511287798066],
+                "center": None,
+                "fillzoom": None,
+                "legend": None,
+                "maxzoom": 30,
+                "scheme": "xyz",
                 "minzoom": 0,
                 "name": "Feature tileset",
                 "tilejson": "3.0.0",
-                "tiles": ["/features/mapbox/tile/{z}/{x}/{y}"],
-                "vector_layers": [
-                    {
-                        "description": "Feature layer",
-                        "fields": {},
-                        "id": "features",
-                        "maxzoom": 22,
-                        "minzoom": 0,
-                    }
-                ],
+                "tiles": ["http://testserver/features/tile/{z}/{x}/{y}"],
+                "vector_layers": [],
+                "version": "1.0.0",
             },
-        )
-
-    def test_postgis_layer(self):
-        self.maxDiff = None
-        response = self.client.get(
-            reverse("layer-postgis-tilejson", args=(self.layer.pk,))
-        )
-        self.assertEqual(response.status_code, 200)
-        content = response.json()
-        self.assertDictEqual(
-            content,
-            {
-                "attribution": "© JEC",
-                "description": "generated from postgis database",
-                "maxzoom": 22,
-                "minzoom": 0,
-                "name": "Layer's features tileset",
-                "tilejson": "3.0.0",
-                "tiles": ["/layer/2/postgis/tile/{z}/{x}/{y}"],
-                "vector_layers": [
-                    {
-                        "description": "Feature layer",
-                        "fields": {},
-                        "id": "features",
-                        "maxzoom": 22,
-                        "minzoom": 0,
-                    }
-                ],
-            },
-        )
-
-    def test_postgis_features(self):
-        self.maxDiff = None
-        response = self.client.get(reverse("feature-postgis-tilejson"))
-        self.assertEqual(response.status_code, 200)
-        content = response.json()
-        self.assertDictEqual(
-            content,
-            {
-                "attribution": "© JEC",
-                "description": "generated from postgis database",
-                "maxzoom": 22,
-                "minzoom": 0,
-                "name": "Feature tileset",
-                "tilejson": "3.0.0",
-                "tiles": ["/features/postgis/tile/{z}/{x}/{y}"],
-                "vector_layers": [
-                    {
-                        "description": "Feature layer",
-                        "fields": {},
-                        "id": "features",
-                        "maxzoom": 22,
-                        "minzoom": 0,
-                    }
-                ],
-            },
-            content,
         )
 
     def test_tilejson_view_default(self):