+++ docs/changelog.rst
@@ -1,8 +1,21 @@
-CHANGELOG
+=========
+Changelog
 =========
 
-1.0.0.a1
---------
+1.0.0-beta
+----------
+
+* Drop python 3.6 and Django 2.2
+* Add python 3.11 and Django 4.2
+
+**Breaking changes**
+
+  * Refactor PostGIS and Python (old named MapBox) backends usage. Use setting to set (default postgis)
+  * No DetailView anymore. As Tile can have many layers, declare VectorLayer on MVTView (one or many).
+
+* Enhancements
+
+  * Add compatibility to use with psycopg v3
 
 * Features